+++ components/src/main/java/com/mercadolibre/android/andesui/message/bodylinks/AndesBodyLinks.kt
@@ -1,0 +1,3 @@
+package com.mercadolibre.android.andesui.message.bodylinks
+
+class AndesBodyLinks(val links: List<AndesBodyLink>, val listener: (index: Int) -> Unit)
